Fixed sleep + added temp reading + inventory check on make

+++ SmartCoffee/lib/TemperatureSensor/TemperatureSensor.h
@@ -5,7 +5,7 @@ class TemperatureSensor {
  
 public:
     TemperatureSensor();
-    float getTemperature();
+    double getTemperature();
 };
 
 #endif
